Very simple ajaxification of the tabs
We load the target document via Ajax and replace our current body
content with it, also replacing the URL if the HTML5 history API is
supported.

The reason to do this is to keep URLs simple and clean for sharing and
in the same time allow someone to share a link to a given view point,
without the link containing a #learn-more anchor, so that the recipient
sees the splash screen.