Fix invalid authenticity token errors
The fix is to force the authenticity_token param to be present in the
form, even though the form is an AJAX one.

These errors would occur when a user visits the site without JavaScript.
The reason for that is the support form's `remote: true` flag was
preventing an authenticity_token to be included in the form fields
because AJAX forms are whitelisted by default. However, when JS is
turned off, the form would not work. Having this param present in the
form does not prevent it from working when JS is enabled.